New calculation approach.

+++ src/geometry/Point.java
@@ -5,10 +5,10 @@ import geometry.*;
 public class Point extends TwoDMO {
 
   public Point() { super(); }
-  public Point(double x, double y) { super(x, y); }
+  public Point(float x, float y) { super(x, y); }
   public Point(TwoDMO Other) { super(Other.x, Other.y); }
 
-  public String getDbg() { return "x: " + Double.toString(x) + "; y: " + Double.toString(y); }
+  public String getDbg() { return "x: " + Float.toString(x) + "; y: " + Float.toString(y); }
 
   public void move(Vector Dir) { x+=Dir.getX(); y+=Dir.getY(); }
   public Point displace(Vector Dir) { return new Point(x+Dir.getX(), y+Dir.getY()); }
